Remove pytest fixtures from utils
